+++ app/source.js
@@ -12,6 +12,14 @@ window.Source = function(templates, cwd, tm, source, revlayers, examples, isMapb
             };
             if (datasource.table) cmParams.value = datasource.table;
             code = CodeMirror($('#layers-' + id + ' div.sql').get(0), cmParams);
+
+            code.setOption('extraKeys', {
+                Tab: function(cm) {
+                    var spaces = Array(cm.getOption('indentUnit') + 1).join(' ');
+                    cm.replaceSelection(spaces);
+                }
+            });
+
             code.getWrapperElement().id = 'layers-' + id + '-code';
         }
         var layer = {
@@ -334,13 +342,14 @@ window.Source = function(templates, cwd, tm, source, revlayers, examples, isMapb
             layers[layer.id] = Layer(layer.id, layer.Datasource);
             orderLayers();
 
-            //set maxzoom, if needed
+            // set maxzoom, if needed
             var maxzoomTarget = $('.max');
             if (maxzoomTarget.val() < metadata.maxzoom) maxzoomTarget.val(metadata.maxzoom);
 
             //show new layer
             var center = metadata.center;
-            map.setView([center[1], center[0]], metadata.maxzoom);
+            var zoom = Math.max(metadata.minzoom, view.model.get('minzoom'));
+            map.setView([center[1], center[0]], zoom);
 
             //open proper modal, depending on if there are multiple layers
             if (layersArray.length > 1) {
@@ -645,7 +654,7 @@ window.Source = function(templates, cwd, tm, source, revlayers, examples, isMapb
     Editor.prototype.zoomToLayer = function(ev) {
         var id = $(ev.currentTarget).attr('id').split('zoom-').pop();
         var filepath = layers[id].get().Datasource.file;
-
+        var view = this;
         // Set map in loading state
         $('#full').addClass('loading');
 
@@ -654,8 +663,8 @@ window.Source = function(templates, cwd, tm, source, revlayers, examples, isMapb
             success: function(metadata) {
                 // Clear loading state
                 $('#full').removeClass('loading');
-                var center = metadata.center,
-                    zoom = Math.max(metadata.minzoom, metadata.maxzoom - 1);
+                var center = metadata.center;
+                var zoom = Math.max(metadata.minzoom, view.model.get('minzoom'));
                 map.setView([center[1], center[0]], zoom);
             },
             error: function(jqXHR, textStatus, errorThrown) {