+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 ### 0.2.6
 - Fix rendering issues with when font list titles are focused.
 - Don't show UTF grid tooltips in xray mode.
-
+- Source UI: Tab key now creates soft tabs (2 spaces) instead of Tab character.
 
 ### 0.2.5
 - Remove Windows atom shell logging as a stopgap fix.